requirements.txt
pandas
matplotlib
firebase-admin
faker
python-dotenv